Extract command gen logic
