+++ src/components/StepCommandPreview.jsx
@@ -1,160 +1,16 @@
 import React, { useState, useEffect } from 'react';
+import { generateCommand } from './commandGenerationUtils';
 
 export default function StepCommandPreview({ data }) {
-  // Get data from previous steps
-  const { providers = [], providerConfigs = {} } = data || {};
-  const generalConfig = data.generalConfig || {};
-  
   // Add command type selection
   const [commandType, setCommandType] = useState('docker');
-
   // Generate command whenever data or command type changes
   const [commandText, setCommandText] = useState('');
   
   useEffect(() => {
-    generateCommand();
+    // Use the extracted utility function
+    setCommandText(generateCommand(data, commandType));
   }, [commandType, data]);
-  
-  const generateCommand = () => {
-    const exportLines = [];
-    const cmdArgs = [];
-    
-    // Process general configuration
-    if (generalConfig && generalConfig.fields) {
-      Object.entries(generalConfig.fields).forEach(([key, config]) => {
-        if (!config || !config.value) return;
-        
-        const { value, asEnv, envKey } = config;
-        
-        if (asEnv) {
-          // Always add exports for environment variables
-          exportLines.push(`export ${envKey}=${value}`);
-          
-          if (commandType === 'docker') {
-            // For Docker, pass env var by name
-            cmdArgs.push(`-e ${envKey}`);
-          }
-        } else {
-          if (commandType === 'docker') {
-            // For Docker, pass as direct env var
-            cmdArgs.push(`-e ${envKey}=${value}`);
-          } else {
-            // For Java, transform to Java format
-            const javaKey = envKey.toLowerCase().replace(/_/g, '.');
-            cmdArgs.push(`--${javaKey}=${value}`);
-          }
-        }
-      });
-    }
-    
-    // Process commit options
-    if (generalConfig && generalConfig.commitOptions && generalConfig.commitOptions.length > 0) {
-      generalConfig.commitOptions.forEach((option, index) => {
-        const envKey = `MODERNE_AGENT_DEFAULTCOMMITOPTIONS_${index}`;
-        
-        if (commandType === 'docker') {
-          cmdArgs.push(`-e ${envKey}=${option}`);
-        } else {
-          const javaKey = `moderne.agent.defaultCommitOptions[${index}]`;
-          cmdArgs.push(`--${javaKey}=${option}`);
-        }
-      });
-    }
-    
-    // Process SCM providers configurations
-    providers.forEach(providerId => {
-      const config = providerConfigs[providerId];
-      if (!config) return;
-      
-      const { instances = [] } = config;
-      
-      instances.forEach((instance, instanceIndex) => {
-        if (!instance) return;
-        
-        Object.entries(instance).forEach(([fieldKey, fieldData]) => {
-          if (!fieldData || !fieldData.value) return;
-          
-          const { value, asEnv, envKey } = fieldData;
-          
-          // Handle array fields differently
-          if (Array.isArray(value)) {
-            value.forEach((item, arrayIndex) => {
-              const arrayEnvKey = envKey.replace(/\${i}/g, arrayIndex);
-              if (asEnv) {
-                exportLines.push(`export ${arrayEnvKey}=${item}`);
-                if (commandType === 'docker') {
-                  cmdArgs.push(`-e ${arrayEnvKey}`);
-                }
-              } else {
-                if (commandType === 'docker') {
-                  cmdArgs.push(`-e ${arrayEnvKey}=${item}`);
-                } else {
-                  const javaArg = transformToJavaFormat(arrayEnvKey, item, instanceIndex);
-                  cmdArgs.push(javaArg);
-                }
-              }
-            });
-          } else {
-            // Handle non-array fields as before
-            if (asEnv) {
-              exportLines.push(`export ${envKey}=${value}`);
-              if (commandType === 'docker') {
-                cmdArgs.push(`-e ${envKey}`);
-              }
-            } else {
-              if (commandType === 'docker') {
-                cmdArgs.push(`-e ${envKey}=${value}`);
-              } else {
-                const javaArg = transformToJavaFormat(envKey, value, instanceIndex);
-                cmdArgs.push(javaArg);
-              }
-            }
-          }
-        });
-      });
-    });
-    
-    // Build the command string
-    let command = '';
-    
-    // Add export lines if any
-    if (exportLines.length > 0) {
-      command += exportLines.join('\n') + '\n\n';
-    }
-    
-    // Add the command based on type
-    if (commandType === 'docker') {
-      command += `docker run ${cmdArgs.join(' ')} -p 8080:8080 moderne-agent:latest`;
-    } else {
-      command += `java -jar moderne-agent-{version}.jar ${cmdArgs.join(' ')}`;
-    }
-    
-    setCommandText(command);
-  };
-  
-  // Convert env var key format to Java argument format
-  const transformToJavaFormat = (envKey, value, index) => {
-    // Example: MODERNE_AGENT_GITHUB_0_URL → moderne.agent.github[0].url
-    let javaKey = envKey.toLowerCase();
-
-    // Replace ${i} placeholders if present
-    javaKey = javaKey.replace(/\${i}/g, '0');
-    
-    // Replace underscores with dots
-    javaKey = javaKey.replace(/_/g, '.');
-    
-    // Handle array indices - match patterns like .0. or .1.
-    javaKey = javaKey.replace(/\.(\d+)\./g, (match, digit) => {
-      return `[${digit}].`;
-    });
-    
-    // Handle the last index if it ends with a digit
-    javaKey = javaKey.replace(/\.(\d+)$/, (match, digit) => {
-      return `[${digit}]`;
-    });
-    
-    return `--${javaKey}=${value}`;
-  };
 
   const copyToClipboard = () => {
     navigator.clipboard.writeText(commandText).then(
@@ -205,6 +61,7 @@ export default function StepCommandPreview({ data }) {
           id="copy-button"
           className="copy-button button button--primary button--sm"
           onClick={copyToClipboard}
+          aria-label="Copy command to clipboard"
         >
           Copy
         </button>
